move make_sqlalchemy_uri helper
